+++ bbot/core/helpers/web/__init__.py
@@ -1,0 +1,1 @@
+from .web import WebHelper
